Seperate the channel from the tag
Slack allows us to use the same webhook url to send to all channels in a given organization. This makes things a lot easier because for each tag and channel, we don't need a new url. Unfortunately, this wouldn't work with our previous structure. This is the quickest move to allow us to use this structure to our benefit. I'll be following up with more refactoring (especially in the webhook notifier class) but I wanted to do as little as possible in the first pass to get it working.