Adds Project with a Project Channel
For client projects, we will likely have a project-specific Slack channel. We would like to have Beggar notificy the project-specific Slack channel of pull requests in addition to any channels that may be tagged in the PR description.

A Project will require a Channel. Channels can be tied to multiple projects. There's no firm requirement these channels be project specific, but that's how I see this being used.

To add a Project Channel and a Project, use Beggar's Active Admin GUI.