Store the tags directly on the pull request
Prior to this commmit, we did not attach tags to a pull request, we attach channels. So let's say that PR is tagged as #ember, we find a channel, with that tag name and then attach the channel to the PR because that was most important association to us. The theory was that we really only cared about the tags for finding the channel because we had to know what channel to post it to.
The problem with that was that when we post it to slack, we wanted to show the tags so people know what they're getting into. Because the PullRequest model only had channels, we done this by grabbing all the channels and outputting the tags for those channels. A channel can have many tags. In our scenario above, your "channel" would be #javascript and that channel would have many other tags associated it with it such as'#backbone and #angular. This made the fact that we included the tags fairly useless because you had no idea which technologies were actually being used.

To resolve this, we're keeping the channels as is but we're going to directly link your tags so that we know exactly what you tagged the PR with.