Use a single webhook URL for posting to Slack
Slack's Incoming Webhooks API makes it easy
to specify a channel to post to by name.

Taking advantage of that feature means we no longer have to set up
different incoming webhook integrations for each channel - we can
use a single incoming webhook, and provide the channel name whenever we
post to Slack.