Add heroku deploy
This adds the setup necessary for the heroku automagical deploy!

Here's the rub: I don't know if it works :cry:. It seems that I can't actually test this without the repo being open. I did validate that the file is correct through heroku. For now, I think we can merge it and as soon as we open it, I'll verify it before we make any announcements.