Rename Beggar to PesteR
    pes·ter
   /ˈpestər/
   _verb_

   trouble or annoy (someone) with frequent or persistent requests or
   interruptions.

   synonyms: badger, hound, harass, plague, annoy, bother, trouble,
             keep after, persecute, torment, bedevil, harry, worry,
             beleaguer, nag, hassle, bug, devil, get on someone's case

   > "is there someplace I can study where no one will pester me?"

Besides the obvious wordplay, PesteR doesn't invoke the idea of someone begging
for help and rather suggests that we're being bothered.

This is more in line with how we consider, for example, Hound: a service to be
annoying about things like style so that nobody has to be 'that person'
troubling coworkers. PesteR pings regularly to remind us to review our friends
PRs so that they can tighten the feedback cycle and we don't forget.